+++ library/src/android/support/v4/view/ViewPager.java
@@ -157,6 +157,17 @@ public class ViewPager extends ViewGroup {
          * @see ViewPager#SCROLL_STATE_SETTLING
          */
         public void onPageScrollStateChanged(int state);
+
+        /**
+         * Called when the selection state changes. Useful for changing the state beyond the scope
+         * of the ViewPager (e.g., displayed menu items) to reflect the page that is currently
+         * selected.
+         *
+         * @param position Position of the page whose state is changing.
+         * @param object Object of the page whose state is changing.
+         * @param selected New selection state, {@code true} indicating the page is selected.
+         */
+        public void onPageSelectedStateChanged(int position, Object object, boolean selected);
     }
 
     /**
@@ -177,6 +188,11 @@ public class ViewPager extends ViewGroup {
 
         @Override
         public void onPageScrollStateChanged(int state) {
+            // This space for rent
+        }
+
+        @Override
+        public void onPageSelectedStateChanged(int position, Object object, boolean selected) {
             // This space for rent
         }
     }
@@ -409,7 +425,9 @@ public class ViewPager extends ViewGroup {
         int lastPos = -1;
         for (int i=0; i<mItems.size(); i++) {
             ItemInfo ii = mItems.get(i);
-            mAdapter.updateSelectedState(ii.object, ii.position == mCurItem);
+            if (mOnPageChangeListener != null) {
+                mOnPageChangeListener.onPageSelectedStateChanged(ii.position, ii.object, ii.position == mCurItem);
+            }
             if ((ii.position < startPos || ii.position > endPos) && !ii.scrolling) {
                 if (DEBUG) Log.i(TAG, "removing: " + ii.position + " @ " + i);
                 mItems.remove(i);
@@ -441,9 +459,9 @@ public class ViewPager extends ViewGroup {
             while (lastPos <= endPos) {
                 if (DEBUG) Log.i(TAG, "appending: " + lastPos);
                 addNewItem(lastPos, -1);
-                if ((lastPos == 0) && (mCurItem == 0)) {
+                if ((lastPos == 0) && (mCurItem == 0) && (mOnPageChangeListener != null)) {
                     //Creating first item for the first time
-                    mAdapter.updateSelectedState(mItems.get(0).object, true);
+                    mOnPageChangeListener.onPageSelectedStateChanged(0, mItems.get(0).object, true);
                 }
                 lastPos++;
             }